fix logic listing, fix img 404, fix depedency for pdf, fix show prospektus
