+++ gatsby-node.js
@@ -1,0 +1,14 @@
+exports.onCreateWebpackConfig = ({ stage, loaders, actions }) => {
+  if (stage === "build-html" || stage === "develop-html") {
+    actions.setWebpackConfig({
+      module: {
+        rules: [
+          {
+            test: /react-read-pdf/,
+            use: loaders.null(),
+          },
+        ],
+      },
+    })
+  }
+}
