+++ project/lottie/upcoming.json
@@ -11,7 +11,7 @@
           "project/images/CAPT/4.webp",
           "project/images/CAPT/5.webp"
         ],
-        "thumbnail": "/images/CAPT/thumbnail.webp",
+        "thumbnail": "project/images/CAPT/thumbnail.webp",
         "description": "<p>Mendukung kebutuhan dalam negeri dan ekspor, Kapal Satria hadir untuk meningkatkan distribusi produksi hasil bumi Indonesia. Kapal Satria terdiri dari dua kapal, yaitu kapal tongkang yang mampu memuat 5.500 ton bahan material dan kapal tunda dengan dua induk mesin bertenaga kuda.</p><p>Kapal tongkang adalah kapal pengangkut kargo pengangkut material berat seperti batubara, bijih nikel, pasir, dan hasil bumi lainnya. Kapal tongkang tidak memiliki mesin penggerak sendiri, sehingga dibutuhkan kapal tunda yang berperan sebagai manuver dan menunda kapal tongkang.</p><p>Tak redup oleh pandemi, kapal tongkang terus berkontribusi pada rantai operasional pertambangan. Bahkan kapal yang ada saat ini tidak mencukupi kebutuhan sehingga terjadi antrian pelanggan yang menjadi peluang bagi Kapal Satria.</p><br /><p>Tertarik jadi pemilik bisnis idaman yang high demand dan low competitor? Tunggu tanggal listingnya.</p><br /><p>Ingatkan saya dengan project ini:</p>"
       }
     }
