+++ src/components/ProjectDetail/detailProjects.scss
